pyproject.toml
[project]
name = "richorm"
version = "0.1.0"
description = "Simple ORM based on Django framework"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.2.1",
    "pyyaml>=6.0.2",
]

[project.scripts]
richorm = "migrate.command:cli"

